+++ front/src/pages/LetterUserListPage.tsx
@@ -1,0 +1,13 @@
+import LetterUserList from '../components/Letter/LetterUserList/LetterUserList';
+import LetterListHeader from '../components/Letter/LetterListHeader/LetterListHeader';
+
+const LetterUserListPage = () => {
+  return (
+    <>
+      <LetterListHeader />
+      <LetterUserList />
+    </>
+  );
+};
+
+export default LetterUserListPage;
