--- front/src/pages/UserLetterListPage.tsx
@@ -1,13 +1,0 @@
-import LetterListHeader from '../components/Letter/LetterListHeader/LetterListHeader';
-import LetterWrapper from '../components/Letter/LetterWrapper/LetterWrapper';
-
-const UserLetterListPage = () => {
-  return (
-    <>
-      <LetterListHeader />
-      <LetterWrapper />
-    </>
-  );
-};
-
-export default UserLetterListPage;
